exceptasm: Use `long` type for exception code

+++ kernel/arch/x86_64/cpu/except.cc
@@ -5,7 +5,7 @@
 
 struct __exception_context_t {
     __exception_jmp_buf_t __exception_state;
-    int __exception_code;
+    long __exception_code;
     __exception_context_t *__next;
 };
 
@@ -17,17 +17,17 @@ __exception_jmp_buf_t *__exception_handler_add(void)
     return &ctx->__exception_state;
 }
 
-int __exception_handler_remove(void)
+long __exception_handler_remove(void)
 {
     __exception_context_t *top = (__exception_context_t *)
             thread_get_exception_top();
-    int code = top->__exception_code;
+    long code = top->__exception_code;
     thread_set_exception_top(top->__next);
     free(top);
     return code;
 }
 
-int __exception_handler_invoke(int exception_code)
+long __exception_handler_invoke(long exception_code)
 {
     __exception_context_t *top = (__exception_context_t *)
             thread_get_exception_top();
